Tab completion support in bash
This is a rough go at implementing tab completion of module names in
bash.

To test it out, source ped_bash_completion.sh and then try completing
ped commands in the same shell. To make it permanent, copy that file
into /etc/bash_completions.d/

If you're happy with this approach, I'll ensure that the file is
packaged and installed with ped, and make an easy way to install it to
the correct location.